RHBRMS-2437: Rules Project Copying Error in B*MS 6.x
