+++ jbpm-designer-backend/src/test/java/org/jbpm/designer/repository/vfs/VFSRepositoryDefaultFileSystemTest.java
@@ -1,124 +1,79 @@
 /*
- * Copyright 2015 JBoss Inc
+ * Copyright 2016 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
-*/
+ */
 
-package org.jbpm.designer.repository;
+package org.jbpm.designer.repository.vfs;
 
+import org.jbpm.designer.repository.Asset;
+import org.jbpm.designer.repository.AssetBuilderFactory;
+import org.jbpm.designer.repository.Directory;
+import org.jbpm.designer.repository.Repository;
 import org.jbpm.designer.repository.filters.FilterByExtension;
-import org.jbpm.designer.repository.filters.FilterByFileName;
 import org.jbpm.designer.repository.impl.AssetBuilder;
-import org.jbpm.designer.repository.vfs.RepositoryDescriptor;
-import org.jbpm.designer.repository.vfs.VFSRepository;
 import org.jbpm.designer.web.profile.impl.JbpmProfileImpl;
-import org.junit.*;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
 import org.uberfire.java.nio.file.NoSuchFileException;
 
 import java.io.File;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.Map;
 
 import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
-public class VFSRepositoryGitFileSystemTest {
-
-    // TODO change it to generic independent path
-    private static final String REPOSITORY_ROOT = "designer-playground";
-    private static final String VFS_REPOSITORY_ROOT = "git://" + REPOSITORY_ROOT;
-    private static final String USERNAME = "guvnorngtestuser1";
-    private static final String PASSWORD = "test1234";
-    private static final String ORIGIN_URL      = "https://github.com/mswiderski/designer-playground.git";
-    private static final String FETCH_COMMAND = "?fetch";
-    private JbpmProfileImpl profile;
-
-    private static String gitLocalClone = System.getProperty("java.io.tmpdir") + File.separator + "git-repo";
-    private static Map<String, String> env = new HashMap<String, String>();
-
-    private static int counter = 0;
-    
-    private RepositoryDescriptor descriptor;
-    private VFSFileSystemProducer producer;
-
-    @BeforeClass
-    public static void prepare() {
-
-        env.put( "username", USERNAME );
-        env.put( "password", PASSWORD );
-        env.put( "origin", ORIGIN_URL );
-        env.put( "fetch.cmd", FETCH_COMMAND );
-        System.setProperty("org.kie.nio.git.dir", gitLocalClone);
-
-    }
-
-    @AfterClass
-    public static void cleanup() {
-        System.clearProperty("org.kie.nio.git.dir");
-    }
+public class VFSRepositoryDefaultFileSystemTest extends RepositoryBaseTest {
 
     @Before
     public void setup() {
+        new File(REPOSITORY_ROOT).mkdir();
         profile = new JbpmProfileImpl();
+
         producer = new VFSFileSystemProducer();
-        env.put("repository.root", VFS_REPOSITORY_ROOT + counter);
+        HashMap<String, String> env = new HashMap<String, String>();
+        env.put("repository.root", VFS_REPOSITORY_ROOT);
         env.put("repository.globaldir", "/global");
         descriptor = producer.produceFileSystem(env);
     }
 
-    private void deleteFiles(File directory) {
-        for (File file : directory.listFiles()) {
-            if (file.isDirectory()) {
-                deleteFiles(file);
-            }
-            file.delete();
-        }
-    }
-
     @After
     public void teardown() {
-        File repo = new File(gitLocalClone);
+        File repo = new File(REPOSITORY_ROOT);
         if(repo.exists()) {
             deleteFiles(repo);
         }
         repo.delete();
-        repo = new File(".niogit");
-        if(repo.exists()) {
-            deleteFiles(repo);
-        }
-        repo.delete();
-        counter++;
     }
-
+    
     @Test
-    public void testListDirectories() {
+    public void testCreateDefaultVFSRepository() {
 
-        Repository repository = new VFSRepository(producer.getIoService());
+
+        Repository repository = new VFSRepository( producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
-        boolean rootFolderExists = repository.directoryExists("/processes");
+        boolean rootFolderExists = repository.directoryExists("/");
         assertTrue(rootFolderExists);
 
-        Collection<Directory> directories = repository.listDirectories("/");
+        Collection<Directory> directories = repository.listDirectories( "/" );
         assertNotNull(directories);
-        assertEquals(3, directories.size());
-
-        directories = repository.listDirectories("/processes");
-        assertNotNull(directories);
-        assertEquals(1, directories.size());
+        assertEquals(0, directories.size());
     }
 
     @Test
     public void testCreateDirectory() {
-
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
         boolean rootFolderExists = repository.directoryExists("/test");
@@ -150,7 +105,7 @@ public class VFSRepositoryGitFileSystemTest {
         rootFolderExists = repository.directoryExists("/test");
         assertTrue(rootFolderExists);
 
-        AssetBuilder builder = AssetBuilderFactory.getAssetBuilder(Asset.AssetType.Byte);
+        AssetBuilder builder = AssetBuilderFactory.getAssetBuilder( Asset.AssetType.Byte );
         builder.content("simple content".getBytes())
                 .type("png")
                 .name("test")
@@ -162,6 +117,38 @@ public class VFSRepositoryGitFileSystemTest {
 
         boolean assetPathShouldNotExists = repository.directoryExists("/test/test.png");
         assertFalse(assetPathShouldNotExists);
+    }
+
+    @Test
+    public void testListDirectories() {
+        Repository repository = new VFSRepository(producer.getIoService());
+        ((VFSRepository)repository).setDescriptor(descriptor);
+        boolean rootFolderExists = repository.directoryExists("/test");
+        assertFalse(rootFolderExists);
+
+        Directory directoryId = repository.createDirectory("/test");
+        assertNotNull(directoryId);
+        directoryId = repository.createDirectory("/test2");
+        assertNotNull(directoryId);
+        directoryId = repository.createDirectory("/test3/nested");
+        assertNotNull(directoryId);
+
+        rootFolderExists = repository.directoryExists("/test");
+        assertTrue(rootFolderExists);
+
+        rootFolderExists = repository.directoryExists("/test2");
+        assertTrue(rootFolderExists);
+
+        rootFolderExists = repository.directoryExists("/test3");
+        assertTrue(rootFolderExists);
+
+        Collection<Directory> directories = repository.listDirectories("/");
+        assertNotNull(directories);
+        assertEquals(3, directories.size());
+
+        directories = repository.listDirectories("/test3");
+        assertNotNull(directories);
+        assertEquals(1, directories.size());
     }
 
     @Test
@@ -214,10 +201,21 @@ public class VFSRepositoryGitFileSystemTest {
 
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
-        boolean rootFolderExists = repository.directoryExists("/processes");
+        boolean rootFolderExists = repository.directoryExists("/");
         assertTrue(rootFolderExists);
 
-        Collection<Asset> assets = repository.listAssets("/processes");
+        Collection<Asset> assets = repository.listAssets("/");
+        assertNotNull(assets);
+        assertEquals(0, assets.size());
+
+        try {
+            new File(REPOSITORY_ROOT + "/" + "test.txt").createNewFile();
+            new File(REPOSITORY_ROOT + "/" + "test.png").createNewFile();
+        } catch (Exception e) {
+
+        }
+
+        assets = repository.listAssets("/");
         assertNotNull(assets);
         assertEquals(2, assets.size());
     }
@@ -227,18 +225,28 @@ public class VFSRepositoryGitFileSystemTest {
 
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
-        boolean rootFolderExists = repository.directoryExists("/processes");
+        boolean rootFolderExists = repository.directoryExists("/");
         assertTrue(rootFolderExists);
 
-        Collection<Asset> assets = repository.listAssets("/processes");
+        Collection<Asset> assets = repository.listAssets("/");
         assertNotNull(assets);
-        assertEquals(2, assets.size());
+        assertEquals(0, assets.size());
+
+        try {
+            new File(REPOSITORY_ROOT + "/" + "test.txt").createNewFile();
+        } catch (Exception e) {
+
+        }
+
+        assets = repository.listAssets("/");
+        assertNotNull(assets);
+        assertEquals(1, assets.size());
         Asset<String> asset = assets.iterator().next();
 
-        assertEquals("bpmn2", asset.getAssetType());
-        assertEquals("BPMN2-ScriptTask.bpmn2", asset.getFullName());
-        assertEquals("BPMN2-ScriptTask", asset.getName());
-        assertEquals("/processes", asset.getAssetLocation());
+        assertEquals("txt", asset.getAssetType());
+        assertEquals("test.txt", asset.getFullName());
+        assertEquals("test", asset.getName());
+        assertEquals("/", asset.getAssetLocation());
     }
 
     @Test
@@ -246,18 +254,29 @@ public class VFSRepositoryGitFileSystemTest {
 
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
-        boolean rootFolderExists = repository.directoryExists("/images");
+        boolean rootFolderExists = repository.directoryExists("/");
         assertTrue(rootFolderExists);
 
-        Collection<Asset> assets = repository.listAssets("/images");
+        Collection<Asset> assets = repository.listAssets("/");
+        assertNotNull(assets);
+        assertEquals(0, assets.size());
+
+        try {
+            new File(REPOSITORY_ROOT + "/" + "test.png").createNewFile();
+        } catch (Exception e) {
+
+        }
+
+        assets = repository.listAssets("/");
         assertNotNull(assets);
         assertEquals(1, assets.size());
-        Asset<String> asset = assets.iterator().next();
+
+        Asset<byte[]> asset = assets.iterator().next();
 
         assertEquals("png", asset.getAssetType());
-        assertEquals("release-process.png", asset.getFullName());
-        assertEquals("release-process", asset.getName());
-        assertEquals("/images", asset.getAssetLocation());
+        assertEquals("test.png", asset.getFullName());
+        assertEquals("test", asset.getName());
+        assertEquals("/", asset.getAssetLocation());
     }
 
     @Test
@@ -265,40 +284,41 @@ public class VFSRepositoryGitFileSystemTest {
 
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
-        boolean rootFolderExists = repository.directoryExists("/processes/nested");
+        boolean rootFolderExists = repository.directoryExists("/");
         assertTrue(rootFolderExists);
 
-        Collection<Asset> assets = repository.listAssets("/processes/nested");
+        Directory directoryId = repository.createDirectory("/test/nested");
+        assertNotNull(directoryId);
+
+        Collection<Asset> assets = repository.listAssets("/test/nested");
+        assertNotNull(assets);
+        assertEquals(0, assets.size());
+
+        try {
+            new File(REPOSITORY_ROOT + "/test/nested/" + "test.txt").createNewFile();
+        } catch (Exception e) {
+
+        }
+
+        assets = repository.listAssets("/test/nested");
         assertNotNull(assets);
         assertEquals(1, assets.size());
         Asset<String> asset = assets.iterator().next();
 
-        assertEquals("bpmn2", asset.getAssetType());
-        assertEquals("BPMN2-UserTask.bpmn2", asset.getFullName());
-        assertEquals("BPMN2-UserTask", asset.getName());
-        assertEquals("/processes/nested", asset.getAssetLocation());
+        assertEquals("txt", asset.getAssetType());
+        assertEquals("test", asset.getName());
+        assertEquals("test.txt", asset.getFullName());
+        assertEquals("/test/nested", asset.getAssetLocation());
     }
 
     @Test
-    public void testLoadAssetFromPath() throws NoSuchFileException {
+    public void testStoreSingleBinaryAsset() throws NoSuchFileException {
 
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
-        Asset<String> asset = repository.loadAssetFromPath("/processes/BPMN2-ScriptTask.bpmn2");
+        boolean rootFolderExists = repository.directoryExists("/");
+        assertTrue(rootFolderExists);
 
-        assertEquals("bpmn2", asset.getAssetType());
-        assertEquals("BPMN2-ScriptTask", asset.getName());
-        assertEquals("BPMN2-ScriptTask.bpmn2", asset.getFullName());
-        assertEquals("/processes", asset.getAssetLocation());
-        assertFalse(asset.getAssetContent()==null);
-        System.out.print(asset.getUniqueId());
-    }
-
-    @Test
-    public void testStoreSingleBinaryAsset() throws NoSuchFileException{
-
-        Repository repository = new VFSRepository(producer.getIoService());
-        ((VFSRepository)repository).setDescriptor(descriptor);
         Collection<Asset> assets = repository.listAssets("/");
         assertNotNull(assets);
         assertEquals(0, assets.size());
@@ -323,38 +343,13 @@ public class VFSRepositoryGitFileSystemTest {
     }
 
     @Test
-    public void testStoreSingleBinaryAssetSpaceInName() throws NoSuchFileException{
-
-        Repository repository = new VFSRepository(producer.getIoService());
-        ((VFSRepository)repository).setDescriptor(descriptor);
-        Collection<Asset> assets = repository.listAssets("/");
-        assertNotNull(assets);
-        assertEquals(0, assets.size());
-
-        AssetBuilder builder = AssetBuilderFactory.getAssetBuilder(Asset.AssetType.Byte);
-        builder.content("simple content".getBytes())
-                .type("png")
-                .name("test asset")
-                .location("/");
-
-        String id = repository.createAsset(builder.getAsset());
-
-        assertNotNull(id);
-
-        Asset<byte[]> asset = repository.loadAsset(id);
-
-        assertEquals("png", asset.getAssetType());
-        assertEquals("test asset", asset.getName());
-        assertEquals("test asset.png", asset.getFullName());
-        assertEquals("/", asset.getAssetLocation());
-        assertFalse(asset.getAssetContent().length == 0);
-    }
-
-    @Test
     public void testStoreSingleTextAsset() throws NoSuchFileException{
 
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
+        boolean rootFolderExists = repository.directoryExists("/");
+        assertTrue(rootFolderExists);
+
         Collection<Asset> assets = repository.listAssets("/");
         assertNotNull(assets);
         assertEquals(0, assets.size());
@@ -383,11 +378,12 @@ public class VFSRepositoryGitFileSystemTest {
 
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
+
+        boolean rootFolderExists = repository.directoryExists("/");
+        assertTrue(rootFolderExists);
+
         Collection<Asset> assets = repository.listAssets("/");
         assertNotNull(assets);
-        for (Asset aset : assets) {
-            System.out.println(aset.getAssetLocation() + " " + aset.getFullName());
-        }
         assertEquals(0, assets.size());
 
         AssetBuilder builder = AssetBuilderFactory.getAssetBuilder(Asset.AssetType.Text);
@@ -419,7 +415,7 @@ public class VFSRepositoryGitFileSystemTest {
         Collection<Asset> foundAsset = repository.listAssetsRecursively("/", new FilterByExtension("bpmn2"));
 
         assertNotNull(foundAsset);
-        assertEquals(4, foundAsset.size());
+        assertEquals(1, foundAsset.size());
     }
 
     @Test
@@ -443,7 +439,7 @@ public class VFSRepositoryGitFileSystemTest {
 
         id = repository.updateAsset(builder.getAsset(), "", "");
 
-        foundAsset = repository.listAssetsRecursively("/", new FilterByFileName("process.bpmn2"));
+        foundAsset = repository.listAssetsRecursively("/", new FilterByExtension("bpmn2"));
 
         assertNotNull(foundAsset);
         assertEquals(1, foundAsset.size());
@@ -509,7 +505,6 @@ public class VFSRepositoryGitFileSystemTest {
 
     }
 
-
     @Test
     public void testCopyAsset() throws NoSuchFileException {
         Repository repository = new VFSRepository(producer.getIoService());
@@ -543,7 +538,6 @@ public class VFSRepositoryGitFileSystemTest {
 
     }
 
-    @Ignore// git based vfs does not yet support move
     @Test
     public void testMoveAsset() throws NoSuchFileException {
         Repository repository = new VFSRepository(producer.getIoService());
@@ -584,7 +578,6 @@ public class VFSRepositoryGitFileSystemTest {
         assertFalse(sourceAssetExists);
     }
 
-    @Ignore// git based vfs does not yet support move
     @Test
     public void testMoveAndRenameAsset() throws NoSuchFileException {
         Repository repository = new VFSRepository(producer.getIoService());
@@ -625,7 +618,6 @@ public class VFSRepositoryGitFileSystemTest {
         assertFalse(sourceAssetExists);
     }
 
-    @Ignore// git based vfs does not yet support move
     @Test
     public void testRenameAsset() throws NoSuchFileException {
         Repository repository = new VFSRepository(producer.getIoService());
@@ -666,8 +658,7 @@ public class VFSRepositoryGitFileSystemTest {
         assertFalse(sourceAssetExists);
     }
 
-    // disabling this test for now
-    // @Test
+    @Test
     public void testCopyDirectory() throws NoSuchFileException {
         Repository repository = new VFSRepository(producer.getIoService());
         ((VFSRepository)repository).setDescriptor(descriptor);
@@ -705,7 +696,6 @@ public class VFSRepositoryGitFileSystemTest {
 
     }
 
-    @Ignore// git based vfs does not yet support move
     @Test
     public void testMoveDirectory() throws NoSuchFileException {
         Repository repository = new VFSRepository(producer.getIoService());
@@ -744,7 +734,6 @@ public class VFSRepositoryGitFileSystemTest {
 
     }
 
-    @Ignore// git based vfs does not yet support move
     @Test
     public void testMoveEmptyDirectory() throws NoSuchFileException {
         Repository repository = new VFSRepository(producer.getIoService());
@@ -764,6 +753,34 @@ public class VFSRepositoryGitFileSystemTest {
 
         boolean movedDirectoryExists = repository.directoryExists("/source");
         assertFalse(movedDirectoryExists);
+        movedDirectoryExists = repository.directoryExists("/target");
+        assertTrue(movedDirectoryExists);
+
+        foundAsset = repository.listAssets("/target", new FilterByExtension("bpmn2"));
+
+        assertNotNull(foundAsset);
+        assertEquals(0, foundAsset.size());
+    }
+
+    @Test
+    public void testCopyEmptyDirectory() throws NoSuchFileException {
+        Repository repository = new VFSRepository(producer.getIoService());
+        ((VFSRepository)repository).setDescriptor(descriptor);
+        Directory sourceDir = repository.createDirectory("/source");
+
+        boolean directoryExists = repository.directoryExists(sourceDir.getLocation()+sourceDir.getName());
+        assertTrue(directoryExists);
+        Collection<Asset> foundAsset = repository.listAssets("/source", new FilterByExtension("bpmn2"));
+
+        assertNotNull(foundAsset);
+        assertEquals(0, foundAsset.size());
+
+
+        boolean copied = repository.copyDirectory("/source", "/target");
+        assertTrue(copied);
+
+        boolean movedDirectoryExists = repository.directoryExists("/source");
+        assertTrue(movedDirectoryExists);
         movedDirectoryExists = repository.directoryExists("/target");
         assertTrue(movedDirectoryExists);
